requirements.txt
Django==1.7.1
dateutils==0.6.6
djangorestframework==2.3.14
psycopg2==2.5.3
python-dateutil==2.2
titlecase==0.7.1
djangorestframework-gis==0.3
uWSGI==2.0.6
django-extended-choices==0.3.0